refactored checkout page into smaller componenets and hooks
