+++ client/src/pages/checkout/types/ProductDetails.tsx
@@ -1,0 +1,9 @@
+// src/shared/types.ts
+export type ProductDetails = {
+    product_id: string;
+    thumbnail: string;
+    title: string;
+    price: string;
+    quantity: number;
+    stock: number;
+};
